requirements.txt
aiofiles==22.1.0
aiosqlite==0.18.0
anyio==3.5.0
appnope==0.1.2
argon2-cffi==21.3.0
argon2-cffi-bindings==21.2.0
arrow==1.2.3
asttokens==2.0.5
attrs==21.4.0
Babel==2.9.1
backcall==0.2.0
beautifulsoup4==4.10.0
bleach==4.1.0
certifi==2021.10.8
cffi==1.15.0
charset-normalizer==2.0.12
cloudpickle==2.0.0
comm==0.1.2
cycler==0.11.0
debugpy==1.5.1
decorator==5.1.1
defusedxml==0.7.1
entrypoints==0.4
executing==0.8.3
fastjsonschema==2.16.3
fonttools==4.31.1
fqdn==1.5.1
gender-guesser==0.4.0
idna==3.3
ipykernel==6.9.2
ipython==8.1.1
ipython-genutils==0.2.0
ipywidgets==7.7.0
isoduration==20.11.0
jedi==0.18.1
Jinja2==3.0.3
joblib==1.1.0
json5==0.9.6
jsonpointer==2.3
jsonschema==4.4.0
jupyter-client==7.1.2
jupyter-core==4.9.2
jupyter-events==0.6.3
jupyter-server==1.15.6
jupyter-ydoc==0.2.3
jupyter_server_fileid==0.8.0
jupyter_server_terminals==0.4.4
jupyter_server_ydoc==0.6.1
jupyterlab==3.3.2
jupyterlab-pygments==0.1.2
jupyterlab-server==2.11.2
jupyterlab-widgets==1.1.0
kiwisolver==1.4.0
llvmlite==0.38.0
MarkupSafe==2.1.1
matplotlib==3.5.1
matplotlib-inline==0.1.3
mistune==0.8.4
nbclassic==0.3.7
nbclient==0.5.13
nbconvert==6.4.4
nbformat==5.2.0
nest-asyncio==1.5.4
notebook==6.4.10
notebook-shim==0.1.0
numba==0.55.1
numpy==1.21.5
packaging==21.3
pandas==1.4.1
pandocfilters==1.5.0
parso==0.8.3
path==16.6.0
patsy==0.5.2
pexpect==4.8.0
pickleshare==0.7.5
Pillow==9.0.1
platformdirs==3.1.1
prometheus-client==0.13.1
prompt-toolkit==3.0.28
psutil==5.9.0
ptyprocess==0.7.0
pure-eval==0.2.2
pycparser==2.21
Pygments==2.11.2
pyparsing==3.0.7
pyrsistent==0.18.1
python-dateutil==2.8.2
python-json-logger==2.0.7
pytz==2022.1
pytz-deprecation-shim==0.1.0.post0
PyYAML==6.0
pyzmq==22.3.0
requests==2.27.1
rfc3339-validator==0.1.4
rfc3986-validator==0.1.1
rpy2==3.4.5
scikit-learn==1.0.2
scipy==1.8.0
seaborn==0.11.2
Send2Trash==1.8.0
shap==0.40.0
six==1.16.0
slicer==0.0.7
sniffio==1.2.0
soupsieve==2.3.1
stack-data==0.2.0
statsmodels==0.13.2
terminado==0.13.3
testpath==0.6.0
threadpoolctl==3.1.0
tinycss2==1.2.1
tomli==2.0.1
tornado==6.1
tqdm==4.63.0
traitlets==5.1.1
tzdata==2022.1
tzlocal==4.1
uri-template==1.2.0
urllib3==1.26.9
wcwidth==0.2.5
webcolors==1.12
webencodings==0.5.1
websocket-client==1.3.1
widgetsnbextension==3.6.0
xgboost==1.5.2
y-py==0.5.9
ypy-websocket==0.8.2
